Add logic to sitecustomize
