+++ Dockerfile.tmpl
@@ -651,10 +651,6 @@ ADD patches/sitecustomize.py /root/.local/lib/python3.10/site-packages/sitecusto
 # Override default imagemagick policies
 ADD patches/imagemagick-policy.xml /etc/ImageMagick-6/policy.xml
 
-# Add generativeai Python patch
-ADD patches/generativeaipatch.py /root/.local/lib/python3.10/site-packages/generativeaipatch.py
-CMD ["python", "/root/.local/lib/python3.10/site-packages/generativeaipatch.py"]
-
 # Add Kaggle module resolver
 ADD patches/kaggle_module_resolver.py /opt/conda/lib/python3.10/site-packages/tensorflow_hub/kaggle_module_resolver.py
 RUN sed -i '/from tensorflow_hub import uncompressed_module_resolver/a from tensorflow_hub import kaggle_module_resolver' /opt/conda/lib/python3.10/site-packages/tensorflow_hub/config.py && \
